Catawiki auction scraper
The following code creates a data frame which for each row collects data representing each bid placed on an item in an auction set, in the case that no bids were placed for an item one row would have the data pertaining to this item. Each auction set represents a collection of items that were up for auction together.

Each auction set at Catawiki has an end time, where each item is differentiated by ending 55 seconds after the item before it in the auction set. To combat sniping Catawiki automatically extends auctions by 90 seconds if there is a bid made in the last minute of the auction. This automated extension will be repeated until there has been no bid placed in the last minute of an individual item’s auction time.
Weekly auction sets run on average for 7-10 days and most categories have minimum one auction set finishing every day. 

This code was used to scrape the data from Classical Art Auctions that took place between the 19th and 27th of May 2021. Data refers to the 2-3 auctions sets that ended per day in this time frame, with a total of 19 auctions sets. 

The following data points are included in the data frame: Auction_id which gives a unique number to each auction set. Auction_id_type which gives a id number depending on the type of classical auction type, types listed in notes of code. Item_title which is named of item. Item_id. favorite_count includes the amount of times that this item has been saved to the favourite list of a member of Catawiki. Bidder and Bidder_id, which is unique for each bidder. Amount has amount of bid placed in euros. Bidded_time which has the time of bid in YYYY-MM-DDTHH:MM:SSZ format. Bid_time_date which has bid date in YYYY-MM-DD format. Bid_time which has bid time in HH:MM:SS format. Start_time which has the time the item’s auction opens in YYYY-MM-DDTHH:MM:SSZ format. Item_start_date which has the time the item’s auction opens in YYYY-MM-DD format. Item_start_time which has the time the item’s auction opens in HH:MM:SS format. Planned_end which has the planned end time of an item’s auction if not extended in YYYY-MM-DDTHH:MM:SSZ format.  Planned_end_time which has the planned end time of an item’s auction if not extended in HH:MM:SS format. 
Item_end_date which has the time the item’s auction ends including possible extended time in YYYY-MM-DD format. Item_end_time which has the time the item’s auction ends including possible extended time  in HH:MM:SS format. Minimum_interval_bid_amount is the minimum amount higher the bid had to be compared to the last bid on that item according to catawiki’s restrictions. Reserve_price_met is a boolean for if at the end of the auction the item’s reserve price was met. Lowest_bid_amount is the lowest bid placed for each item. Highest_bid_amount is the highest bid placed for each item. Total_bids are the amount of bids that were placed for each item. Url has the url for each item’s auction. auction_url is the url for the auction set in which the item was. Auction_name is the name displayed for each auction set. Number_of_bidders is the number of unique bidders who placed bids for each auction. Auction_extended is a binary variable which is 1 if the auction was automatically extended once or more.